Added special case for using rolling operation with date dimension and date filter in order to adjust the filter then truncate the resulting data frame
